requirements.txt
asgiref==3.7.2
Django==4.2.7
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pycodestyle==2.11.1
pytest==7.4.3
sqlparse==0.4.4
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-bootstrap5==23.3
django-crispy-forms==2.1
django-filter==23.3
django-tables2==2.6.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pycodestyle==2.11.1
pytest==7.4.3
PyYAML==6.0.1
sqlparse==0.4.4
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-bootstrap5==23.3
django-crispy-forms==2.1
django-filter==23.3
django-tables2==2.6.0
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pytest==7.4.3
PyYAML==6.0.1
sqlparse==0.4.4
